[boringssl] Port GN files from build/secondary
This is the first buildable  commit for the new BoringSSL tree.
Originally, Fuchsia forked BoringSSL to modify it to support the OS.
Over time, these small modifcations have been upstreamed, allowing
BoringSSL to become a pure mirror of upstream, with build files under
//build using GN's secondary_source feature.  Unfortunately, this
doesn't work well with the "layer cake" model, as it the source and
build files cannot be atomically updated.  The solution is to use
//third_party/boringssl as an extra layer with the build files and a
jiri manifest that directly references upstream.  This is much closer
to Chromium's approach, and has more support in the src/util/ scripts.

Change-Id: I3c5d932f9526b21df07eccfaddbc996fbb41c8db