[gn] Clean up GN, remove libcrypto package
Fuchsia's BUILDCONFIG.gn now provides default_library_type, so use it.

Library-only packages should not exist and libcrypto has no reason to
be a package, so remove it.  The libssl package is still necessary
just for its resources, but remove the actual library from the package
specification so it's found implicit by dependencies.

Change-Id: I7d10921fdf45fa35da8cbbdcd9ad1ff74ddad444