[fuchsia] Link against zxcpp
Previously, we were linking boringssl dynamically against libc++.so.
Because we include boringssl in the SDK, we want to not link against
libc++.so.

In order for the C++ language to work, we need to link against some
library that provides a few basic symbols. This CL obtains those symbols
by statically linking against zxcpp, which is how we normally get those
symbols without libc++.

Test: SDK compiles
Change-Id: I43e4220a62ea1baae1a172f21b781a496842db44