Refactor rollers for fuchsia.git, integration repo
This CL updates the roller scripts to incorporate two major changes to
the Fuchsia project:
 1. Jiri manifests have been moved to an internal integration repo.
 2. The "stem", including zircon and garnet has been made monorepo.

The result of both of these changes is that roll_boringssl.go script can
no longer update manifests, and thus no longer needs to make commits
automatically (which was previously error-prone).  Updating the internal
manifests is now a manual step. A 'check-integration' script is added to
make it simple, and the README.fuchsia has been updated with current
instructions.

Change-Id: I906ce9da89758a2e05887da54cd96dd7078a8c44