[rust] Pin to a particular version of bindgen
Changes to bindgen sometimes change the semantics of the generated
bindings (for example, by changing the Rust types that are generated
for particular C types). This change modifies `bindgen.sh` to "pin"
to a particular version of bindgen so that developers do not
accidentally use an older version of bindgen. Developers may still
use a more recent version of bindgen when one is available by editing
`bindgen.sh` to pin that more recent version.

Change-Id: I16a74c35e8eee8cf956c7e6867fbeb59e4d48bf7
Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/boringssl/+/376416
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Tamir Duberstein <tamird@google.com>